fix: Filter out duplicate events by tracking event keys seen
Because the linearized core is truncated and blocks are shuffled around
and hence replay in the hyperbee history stream, a mechanism is needed
to filter out duplicate emits of a single event. Since keys are unique
to the event itself, the keys of emitted events are tracked and checked
for subsequent attempts to emit.

Ideally I'd like a history stream that didn't reemit key/values but
given that the linearized core is where the hyperbee is stored we can't
use the key's presence as a filter. It's effectively rolled back to a
checkpoint on truncation of the linearized core.