refact: Convert event emitting to use diff streams
Now the event emitting scan is fired by append events on the view log
when caught up. Because events can be appended as a result of emitting
other events, a check for immediate rescan it fired if hyperbee versions
don't match.

Truncation is also accounted for assuming that the event emitting scan
needs to start from the sequence number that the view log was truncated
to. This would cause double firing of events if there wasn't the seen
set.

Tests include more exhaustive testing including replication testing.