fix: Add `key` index key to ensure no double fire
With the new input core key & sequence filter for filtering events
already fired, there was an edge case that could skip events. If an
event was later in the input key sequence but would show up in the diff
stream earlier because the event name was earlier alphabetically, then
its sequence number would cause events alphabetically later but
sequentially before to not be fired.

I attempted to use the `time` index instead which works as long as the
two events in question do not have the same timestamp. I'm sure there
could the analogous edge case though for events entered with reverse
order timestamps and sequence numbers.

So to solve this, I now only rely on the input key and sequence number
for firing all events in diff stream. This has some drawbacks:

1. Events cannot be ordered by event or timestamp for when they fire.
   This is not too bad as instead they are just fired in sequential
   order per input core which makes intuitive sense but might be weird
   with events firing out of order on different input cores.
2. Adding yet another index which copies all the event data. This might
   be fixed in the future via storing this new `key` key which is
   guaranteed to be unique per event as the value of other indexes. This
   only saves space assuming that the event object is smaller than the
   key, which is likely.