feat: Add RangeWatcher class to support a hyperbee range per on listener
Creating a diff stream per event allows for less iteration through the
index theoretically. It does require the secondary index for preventing
double emits to be created per watcher.

The `_started` promise for guarding creating watchers until the index is
first created feels gross. I'm sure there is a better way to implement
this guard than a promise that resolves on an event trigger.

Finally, events now can be triggered before the `append` event on the
`autobase`. Instead it can be setup as early as the `autobase` is
`ready`.