chore: Clean up index formatting
Decided that setting `eagerUpdate` default option passed to `autobase`
didn't make sense. The options for `EventBus` should be as similar to
`autobase` as possible to make use of its approach and making use of
`EventBus` just a layer on top of `autobase`. Currently the logic in
`autobase` looks at whether a `localOutput` is defined to determine
whether to `eagerUpdate` or not.