fix: Update view on start even w/ localOutput
The view core does not have the correct length without the update
causing the current view db version to be `1` aka from the beginning.
This means that the entirety of the event stream is emitted when a
localOutput is present.

Eventually this should be an option to allow this as a configurable
feature. The default should be to skip to the last known event as this
is the more common use case matching the how EventEmitter works being
not persistent.