fix: Allow view to be truncated mid diff stream
The `truncate` listener is now set at the beginning of the diff stream
so that if truncation happens while iterating through the stream it will
stop the current iteration, destroy the stream and updates the
`this.latestDiff` so next time around it start the stream from the point
it was truncated.

Needed to add a condition to the normal update of `this.latestDiff` so
that the update from the truncate hook doesn't get overwritten.

Finally the diff stream is immediately restarted if it was truncated
during iterating.